Wrote an untested faceInfo class and implemented it in utils
FaceInfo objects store distances that are normalized based on the
distance between the eyes used to compared faces in a dictionary for
easy conversion to and from JSON. FaceInfo has a class variable for
percent tolerance, and an isSamePerson(other) method for comparing
faces.

The analyze_photo method in utils now creates a face object, has it
generate the values from a stasm landmarks array, and returns the info
dictionary from the face.